+++ src/encoding/evm/strategy_encoder/mod.rs
@@ -1,2 +1,2 @@
+pub mod strategy_encoder_registry;
 mod strategy_encoders;
-pub mod strategy_selector;