+++ src/encoding/mod.rs
@@ -2,6 +2,6 @@ mod errors;
 #[cfg(feature = "evm")]
 pub mod evm;
 pub mod models;
-mod strategy_encoder;
+pub mod strategy_encoder;
 mod swap_encoder;
 pub mod tycho_encoder;